Consolidating provisioning metaports functionality in a single service
