+++ provisioning-api/src/ioc/__init__.py
@@ -16,8 +16,7 @@ from src.provisioner.image_runner import DockerEngineImageRunner
 from src.provisioner.provisioner import Provisioner
 from src.services.provisioning_tasks import ProvisioningTaskService
 from src.services.domains import DomainService
-from src.services.provisioning_metaports_mappings import ProvisioningMetaportsMappingService
-from src.services.provisioning_metaports import ProvisioningMetaportsService
+from src.services.metaports import MetaportsService
 
 class ProvisionerContainer(DeclarativeContainer):  # pylint: disable=too-few-public-methods
   """Provides all the bindings required by the provisioner service."""
@@ -48,12 +47,8 @@ class ProvisionerContainer(DeclarativeContainer):  # pylint: disable=too-few-pub
                              conn_pool=hbase_conn_pool,
                              charset=CoreContainer.DEFAULT_CHARSET)
 
-  metaports_mapping_service = Singleton(ProvisioningMetaportsMappingService,
-                                        flask_app=CoreContainer.config.flask_app,
-                                        conn_pool=hbase_conn_pool,
-                                        charset=CoreContainer.DEFAULT_CHARSET)
 
-  metaports_service = Singleton(ProvisioningMetaportsService,
+  metaports_service = Singleton(MetaportsService,
                                 flask_app=CoreContainer.config.flask_app,
                                 conn_pool=hbase_conn_pool,
                                 charset=CoreContainer.DEFAULT_CHARSET)
@@ -65,7 +60,6 @@ class ProvisionerContainer(DeclarativeContainer):  # pylint: disable=too-few-pub
   docker_image_runner = Singleton(DockerEngineImageRunner, hbase_conn_pool=hbase_conn_pool,
                                   default_charset=CoreContainer.DEFAULT_CHARSET,
                                   docker_client=docker_client,
-                                  metaports_mapping_service=metaports_mapping_service,
                                   metaports_service=metaports_service)
 
   provisioner = Singleton(Provisioner,
